requirements.txt
Django==5.2.4
gunicorn
whitenoise
dj-database-url
psycopg[binary]
djangorestframework
requests
